+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## in develop
 
 * `TAT.Workflow` has a `source` attribute in analogy to `TAT.Document` to be used for checksum calculation for App/Job reuse
+* Fixes evaluation of structs as input parameters of workflow/scatter: hash inputs are coerced from object to struct, and their optional elements are assigned to Null if not specified in job inputs.
 
 ## 0.17.8 (2022-03-15)